Move all the implementation details for Xgit.Repository.OnDisk into a single file.
